add option to skip download

+++ scripts/prepare.sh
@@ -4,7 +4,7 @@
 # get command line arguments
 version=${1}
 external_id=${2:-petsc}
-
+clean=${3:-clean}
 # prepare necessary variables
 url="https://web.cels.anl.gov/projects/petsc/download/release-snapshots/petsc-$version.tar.gz"
 dir="ocp/external/$external_id/$version/source"
@@ -14,9 +14,12 @@ gz_file="$external_id-v$version.tar.gz"
 cd $script_dir/..
 
 # start the script
-rm -rf *.tar.gz
-wget -O $gz_file $url
-echo "Download $external_id $version"
+# if the flag is set, remove the previous tar.gz file
+if [ "$clean" == "clean" ]; then
+    rm -rf *.tar.gz
+    wget -O $gz_file $url
+    echo "Download $external_id $version"
+fi
 rm -rf ocp
 echo "Remove previous ocp folder"
 mkdir -p "$dir"
